add bundled uploads to nft.storage uploader (#2060)
This replaces the nft.storage uploader with one that supports "bundles"
of multiple NFTs. This greatly improves performance by avoiding rate
limits and requiring fewer signatures to authorize uploads.

See https://github.com/nftstorage/metaplex-auth/pull/39 for context.

The existing `batchSize` config param is used to determine how many NFTs
to include in each bundle, with a default of 50 if no `batchSize` is
set, and a maximum of NFTBundle.MAX_ENTRIES (currently 2200).

This PR also adds an `nftStorageGateway` config option that can override
the default HTTP gateway host of `https://nftstorage.link` used in
metadata links, in case the creator has a preferred gateway.

Finally, the `files` array of the final metadata now contains two
entries for each asset file. The first entry, with `cdn = true` is an HTTP
gateway link, while the second, with `cdn = false` is an `ipfs://` URI
that can be used to locate the data using any IPFS peer or gateway. This
should add a bit of resiliency in case the primary gateway should go
down, although that will require clients that know to look for the
alternative `ipfs://` URIs.